chore: split out components
